Allow withParsedBody to accept scalar values
`ServerRequestInterface::withParsedBody` says
> The data IS NOT REQUIRED to come from $_POST, but MUST be the results of
> deserializing the request body content. Deserialization/parsing returns
> structured data, and, as such, this method ONLY accepts arrays or objects,
> or a null value if nothing was available to parse.
> As an example, if content negotiation determines that the request data
> is a JSON payload, this method could be used to create a request
> instance with the deserialized parameters.

A JSON body payload can also be a json string, a json int etc. So withParsedBody would also need to accept that as well. Anybody remember why this is limited to array or object? Those two sentences contradict each other.

According to a slack discussion, this was based on rfc4627

> A JSON text is a serialized object or array.

But that is outdated since is outdated since rfc7158 in 2013 which does not limit it to array or object anymore. See current https://tools.ietf.org/html/rfc8259

> A JSON text is a serialized value.  Note that certain previous
> specifications of JSON constrained a JSON text to be an object or an
> array.